[1/2] cpu_input_boost: Rework display state awareness logic for OnePlus 7 devices
This amends commit cadca83 ("cpu_input_boost: Update SCREEN_AWAKE bit conditional")

Currently, cpu_input_boost fails to engage when the device is
initially powered on until the display is manually cycled off/on.
Additionally, it occasionally fails to re-engage following
subsequent display power cycles. In order to ensure proper
functionality, we need to be more explicit about which display
states the driver is checking for.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: LibXZR <xzr467706992@163.com>